+++ contracts/solutions/day_4/tests/abi/mod.rs
@@ -4,7 +4,7 @@ use e2e::alloy::sol;
 sol!(
     #[sol(rpc)]
     interface Solution {
-      function solvepart1(string calldata input) external view returns (uint32 result);
-      function solvepart2(string calldata input) external view returns (uint32 result);
+      function solvepart1(string calldata input) external view returns (int64 result);
+      function solvepart2(string calldata input) external view returns (int64 result);
   }
 );
